ticket price changed from int to double

+++ swiftyticketing/src/main/java/com/swiftyticket/dto/zone/ZoneRequest.java
@@ -15,5 +15,5 @@ public class ZoneRequest {
     Integer zoneCapacity;
     String zoneName;
     Date zoneDate;
-    int ticketPrice;
+    double ticketPrice;
 }
